+++ api/src/main/java/net/jqwik/api/arbitraries/DoubleArbitrary.java
@@ -13,21 +13,55 @@ import static org.apiguardian.api.API.Status.*;
 public interface DoubleArbitrary extends Arbitrary<Double> {
 
 	/**
-	 * Set the allowed lower {@code min} (included) and upper {@code max} (included) bounder of generated numbers.
+	 * Set the allowed lower {@code min} (included) and upper {@code max} (included) border of generated numbers.
+	 *
+	 * @param min The lower border of possible values
+	 * @param max The upper border of possible values
 	 */
 	default DoubleArbitrary between(double min, double max) {
-		return greaterOrEqual(min).lessOrEqual(max);
+		return between(min, true, max, true);
 	}
 
 	/**
-	 * Set the allowed lower {@code min} (included) bounder of generated numbers.
+	 * Set the allowed lower {@code min} (included) and upper {@code max} (included) border of generated numbers.
+	 * Specify if borders should be included in allowed values or not.
+	 *
+	 * @param min         The lower border of possible values
+	 * @param minIncluded Should the lower border be included
+	 * @param max         The upper border of possible values
+	 * @param maxIncluded Should the upper border be included
+	 */
+	DoubleArbitrary between(double min, boolean minIncluded, double max, boolean maxIncluded);
+
+	/**
+	 * Set the allowed lower {@code min} (included) border of generated numbers.
+	 *
+	 * @param min The lower border of possible values
 	 */
 	DoubleArbitrary greaterOrEqual(double min);
 
 	/**
-	 * Set the allowed upper {@code max} (included) bounder of generated numbers.
+	 * Set the allowed lower {@code min} (excluded) border of generated numbers.
+	 *
+	 * @param min The lower border of possible values
+	 */
+	@API(status = MAINTAINED, since = "1.2.7")
+	DoubleArbitrary greaterThan(double min);
+
+	/**
+	 * Set the allowed upper {@code max} (included) border of generated numbers.
+	 *
+	 * @param max The upper border of possible values
 	 */
 	DoubleArbitrary lessOrEqual(double max);
+
+	/**
+	 * Set the allowed upper {@code max} (excluded) border of generated numbers.
+	 *
+	 * @param max The upper border of possible values
+	 */
+	@API(status = MAINTAINED, since = "1.2.7")
+	DoubleArbitrary lessThan(double max);
 
 	/**
 	 * Set the scale (maximum number of decimal places) to {@code scale}.
@@ -36,6 +70,8 @@ public interface DoubleArbitrary extends Arbitrary<Double> {
 
 	/**
 	 * Set shrinking target to {@code target} which must be between the allowed bounds.
+	 *
+	 * @param target The value which is considered to be the most simple value for shrinking
 	 */
 	@API(status = EXPERIMENTAL, since = "1.1.5")
 	DoubleArbitrary shrinkTowards(double target);